pyproject.toml
[tool.poetry]
name = "tigris-integration"
version = "0.1.0"
description = ""
authors = ["Adil Ansari <adil@tigrisdata.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
tigrisdb = "^1.0.0.b3"
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
